requirements.txt
openai
markdown2
python-dotenv
